ver: Cleanup ifu_mem_ctl test suite
